Remove separate constants
