+++ src/constants/actions.ts
@@ -2,3 +2,9 @@ export const INCREASE_COUNTER = 'INCREASE_COUNTER';
 export const INCREASE_AMOUNT = 'INCREASE_AMOUNT';
 export const DECREASE_COUNTER = 'DECREASE_COUNTER';
 export const INCREASE_ASYNC = 'INCREASE_ASYNC';
+
+export const GET_TODOS = 'GET_TODOS';
+export const ADD_TODOS = 'ADD_TODOS';
+export const UPDATE_TODO = 'UPDATE_TODO';
+export const ADD_TODO = 'ADD_TODO';
+export const STORE_TODO = 'STORE_TODO';
