--- src/constants/todo.ts
@@ -1,5 +1,0 @@
-export const GET_TODOS = 'GET_TODOS';
-export const ADD_TODOS = 'ADD_TODOS';
-export const UPDATE_TODO = 'UPDATE_TODO';
-export const ADD_TODO = 'ADD_TODO';
-export const STORE_TODO = 'STORE_TODO';
